+++ lib/msal-browser/test/app/PublicClientApplication.spec.ts
@@ -7,9 +7,9 @@ const expect = chai.expect;
 import sinon from "sinon";
 import { PublicClientApplication } from "../../src/app/PublicClientApplication";
 import { TEST_CONFIG, TEST_URIS, TEST_HASHES, TEST_TOKENS, TEST_DATA_CLIENT_INFO, TEST_TOKEN_LIFETIMES, RANDOM_TEST_GUID, DEFAULT_OPENID_CONFIG_RESPONSE, testNavUrl, testLogoutUrl, TEST_STATE_VALUES, testNavUrlNoRequest } from "../utils/StringConstants";
-import { ServerError, Constants, AccountInfo, IdTokenClaims, PromptValue, AuthenticationResult, AuthorizationCodeRequest, AuthorizationUrlRequest, IdToken, PersistentCacheKeys, SilentFlowRequest, CacheSchemaType, TimeUtils, AuthorizationCodeClient, ResponseMode, SilentFlowClient, TrustedAuthority, EndSessionRequest, CloudDiscoveryMetadata, AccountEntity, ProtocolUtils } from "@azure/msal-common";
+import { ServerError, Constants, AccountInfo, IdTokenClaims, PromptValue, AuthenticationResult, AuthorizationCodeRequest, AuthorizationUrlRequest, IdToken, PersistentCacheKeys, SilentFlowRequest, CacheSchemaType, TimeUtils, AuthorizationCodeClient, ResponseMode, SilentFlowClient, TrustedAuthority, EndSessionRequest, CloudDiscoveryMetadata, AccountEntity, ProtocolUtils, ServerTelemetryCacheValue } from "@azure/msal-common";
 import { BrowserUtils } from "../../src/utils/BrowserUtils";
-import { BrowserConstants, TemporaryCacheKeys } from "../../src/utils/BrowserConstants";
+import { BrowserConstants, TemporaryCacheKeys, ApiId } from "../../src/utils/BrowserConstants";
 import { Base64Encode } from "../../src/encode/Base64Encode";
 import { XhrClient } from "../../src/network/XhrClient";
 import { BrowserAuthErrorMessage, BrowserAuthError } from "../../src/error/BrowserAuthError";
@@ -58,7 +58,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             done();
         });
 
-        it("navigates and caches hash if navigateToLoginRequestUri is true", (done) => {
+        it("navigates and caches hash if navigateToLoginRequestUri is true and interaction type is redirect", (done) => {
             window.location.hash = TEST_HASHES.TEST_SUCCESS_CODE_HASH;
             window.sessionStorage.setItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`, TEST_URIS.TEST_ALTERNATE_REDIR_URI);
             sinon.stub(BrowserUtils, "navigateWindow").callsFake((urlNavigate: string, noHistory?: boolean) => {
@@ -67,18 +67,19 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 done();
             });
             pca.handleRedirectPromise();
-            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
+            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(null);
         });
 
         it("navigates to root and caches hash if navigateToLoginRequestUri is true", (done) => {
             window.location.hash = TEST_HASHES.TEST_SUCCESS_CODE_HASH;
             sinon.stub(BrowserUtils, "navigateWindow").callsFake((urlNavigate: string, noHistory?: boolean) => {
                 expect(noHistory).to.be.true;
-                expect(urlNavigate).to.be.eq("/");
+                expect(urlNavigate).to.be.eq("https://localhost:8081/");
+                expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`)).to.be.eq("https://localhost:8081/");
                 done();
             });
             pca.handleRedirectPromise();
-            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
+            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(null);
         });
 
         it("navigates to root and caches hash if navigateToLoginRequestUri is true and loginRequestUrl is 'null'", (done) => {
@@ -86,11 +87,12 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             window.sessionStorage.setItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`, "null");
             sinon.stub(BrowserUtils, "navigateWindow").callsFake((urlNavigate: string, noHistory?: boolean) => {
                 expect(noHistory).to.be.true;
-                expect(urlNavigate).to.be.eq("/");
+                expect(urlNavigate).to.be.eq("https://localhost:8081/");
+                expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`)).to.be.eq("https://localhost:8081/");
                 done();
             });
             pca.handleRedirectPromise();
-            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
+            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(null);
         });
 
         it("navigates and caches hash if navigateToLoginRequestUri is true and loginRequestUrl contains query string", (done) => {
@@ -103,7 +105,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 done();
             });
             pca.handleRedirectPromise();
-            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
+            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(null);
         });
 
         it("navigates and caches hash if navigateToLoginRequestUri is true and loginRequestUrl contains query string and hash", (done) => {
@@ -116,7 +118,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 done();
             });
             pca.handleRedirectPromise();
-            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
+            expect(window.sessionStorage.getItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.URL_HASH}`)).to.be.eq(null);
         });
 
         it("replaces custom hash if navigateToLoginRequestUri is true and loginRequestUrl contains custom hash", (done) => {
@@ -125,6 +127,17 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             window.sessionStorage.setItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`, loginRequestUrl);
             sinon.stub(PublicClientApplication.prototype, <any>"handleHash").callsFake((responseHash) => {
                 expect(window.location.href).to.be.eq(loginRequestUrl);
+                expect(responseHash).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
+                done();
+            });
+            pca.handleRedirectPromise();
+        });
+
+        it("processes hash if navigateToLoginRequestUri is true and loginRequestUrl contains trailing slash", (done) => {
+            const loginRequestUrl = window.location.href.endsWith('/') ? window.location.href.slice(0, -1) : window.location.href + "/";
+            window.location.hash = TEST_HASHES.TEST_SUCCESS_CODE_HASH;
+            window.sessionStorage.setItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`, loginRequestUrl);
+            sinon.stub(PublicClientApplication.prototype, <any>"handleHash").callsFake((responseHash) => {
                 expect(responseHash).to.be.eq(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
                 done();
             });
@@ -439,6 +452,17 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 				await expect(pca.loginRedirect(null)).to.be.rejectedWith(BrowserAuthError);
             });
 
+            it("Uses default request if no request provided", (done) => {
+                sinon.stub(pca, "acquireTokenRedirect").callsFake((request) => {
+                    expect(request.scopes).to.contain("openid");
+                    expect(request.scopes).to.contain("profile");
+                    done();
+                    return null;
+                });
+
+                pca.loginRedirect();
+            });
+
             it("loginRedirect navigates to created login url", (done) => {
                 sinon.stub(RedirectHandler.prototype, "initiateAuthRequest").callsFake((navigateUrl): Window => {
                     expect(navigateUrl).to.be.eq(testNavUrl);
@@ -473,7 +497,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
                 sinon.stub(TimeUtils, "nowSeconds").returns(TEST_STATE_VALUES.TEST_TIMESTAMP);
 
-                pca.loginRedirect();
+                pca.loginRedirect(null);
             });
 
 			it("Updates cache entries correctly", async () => {
@@ -543,11 +567,24 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 					challenge: TEST_CONFIG.TEST_CHALLENGE,
 					verifier: TEST_CONFIG.TEST_VERIFIER
 				});
-				const loginUrlErr = "loginUrlErr";
-				sinon.stub(AuthorizationCodeClient.prototype, "getAuthCodeUrl").throws(new BrowserAuthError(loginUrlErr));
-				await expect(pca.loginRedirect(emptyRequest)).to.be.rejectedWith(loginUrlErr);
-				await expect(pca.loginRedirect(emptyRequest)).to.be.rejectedWith(BrowserAuthError);
-				expect(browserStorage.getKeys()).to.be.empty;
+				
+				const testError = {
+                    errorCode: "create_login_url_error",
+                    errorMessage: "Error in creating a login url"
+                }
+                sinon.stub(AuthorizationCodeClient.prototype, "getAuthCodeUrl").throws(testError);
+                try {
+                    await pca.loginRedirect(emptyRequest);
+                } catch (e) {
+                    // Test that error was cached for telemetry purposes and then thrown
+                    expect(window.sessionStorage).to.be.length(1);
+                    const failures = window.sessionStorage.getItem(`server-telemetry-${TEST_CONFIG.MSAL_CLIENT_ID}`);
+                    const failureObj = JSON.parse(failures) as ServerTelemetryCacheValue;
+                    expect(failureObj.failedRequests).to.be.length(2);
+                    expect(failureObj.failedRequests[0]).to.eq(ApiId.acquireTokenRedirect);
+                    expect(failureObj.errors[0]).to.eq(testError.errorCode);
+                    expect(e).to.be.eq(testError);
+                }
 			});
 
 			it("Uses adal token from cache if it is present.", async () => {
@@ -734,12 +771,25 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 				sinon.stub(CryptoOps.prototype, "generatePkceCodes").resolves({
 					challenge: TEST_CONFIG.TEST_CHALLENGE,
 					verifier: TEST_CONFIG.TEST_VERIFIER
-				});
-				const loginUrlErr = "loginUrlErr";
-				sinon.stub(AuthorizationCodeClient.prototype, "getAuthCodeUrl").throws(new BrowserAuthError(loginUrlErr));
-				await expect(pca.acquireTokenRedirect(emptyRequest)).to.be.rejectedWith(loginUrlErr);
-				await expect(pca.acquireTokenRedirect(emptyRequest)).to.be.rejectedWith(BrowserAuthError);
-				expect(browserStorage.getKeys()).to.be.empty;
+                });
+                
+				const testError = {
+                    errorCode: "create_login_url_error",
+                    errorMessage: "Error in creating a login url"
+                }
+                sinon.stub(AuthorizationCodeClient.prototype, "getAuthCodeUrl").throws(testError);
+                try {
+                    await pca.acquireTokenRedirect(emptyRequest);
+                } catch (e) {
+                    // Test that error was cached for telemetry purposes and then thrown
+                    expect(window.sessionStorage).to.be.length(1);
+                    const failures = window.sessionStorage.getItem(`server-telemetry-${TEST_CONFIG.MSAL_CLIENT_ID}`);
+                    const failureObj = JSON.parse(failures) as ServerTelemetryCacheValue;
+                    expect(failureObj.failedRequests).to.be.length(2);
+                    expect(failureObj.failedRequests[0]).to.eq(ApiId.acquireTokenRedirect);
+                    expect(failureObj.errors[0]).to.eq(testError.errorCode);
+                    expect(e).to.be.eq(testError);
+                }
 			});
 
 			it("Uses adal token from cache if it is present.", async () => {
@@ -850,6 +900,17 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 await expect(pca.loginPopup(null)).to.be.rejectedWith(BrowserAuthError);
             });
 
+            it("Uses default request if no request provided", (done) => {
+                sinon.stub(pca, "acquireTokenPopup").callsFake((request) => {
+                    expect(request.scopes).to.contain("openid");
+                    expect(request.scopes).to.contain("profile");
+                    done();
+                    return null;
+                });
+
+                pca.loginPopup();
+            });
+
             it("resolves the response successfully", async () => {
                 const testServerTokenResponse = {
                     token_type: TEST_CONFIG.TOKEN_TYPE_BEARER,
@@ -904,7 +965,10 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             });
 
             it("catches error and cleans cache before rethrowing", async () => {
-                const testError = "Error in creating a login url";
+                const testError = {
+                    errorCode: "create_login_url_error",
+                    errorMessage: "Error in creating a login url"
+                }
                 sinon.stub(AuthorizationCodeClient.prototype, "getAuthCodeUrl").resolves(testNavUrl);
 				sinon.stub(PopupHandler.prototype, "initiateAuthRequest").throws(testError);
 				sinon.stub(CryptoOps.prototype, "generatePkceCodes").resolves({
@@ -915,8 +979,14 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 try {
                     const tokenResp = await pca.loginPopup(null);
                 } catch (e) {
-                    expect(window.sessionStorage).to.be.empty;
-                    expect(`${e}`).to.be.eq(testError);
+                    // Test that error was cached for telemetry purposes and then thrown
+                    expect(window.sessionStorage).to.be.length(1);
+                    const failures = window.sessionStorage.getItem(`server-telemetry-${TEST_CONFIG.MSAL_CLIENT_ID}`);
+                    const failureObj = JSON.parse(failures) as ServerTelemetryCacheValue;
+                    expect(failureObj.failedRequests).to.be.length(2);
+                    expect(failureObj.failedRequests[0]).to.eq(ApiId.acquireTokenPopup);
+                    expect(failureObj.errors[0]).to.eq(testError.errorCode);
+                    expect(e).to.be.eq(testError);
                 }
             });
         });
@@ -992,7 +1062,10 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             });
 
             it("catches error and cleans cache before rethrowing", async () => {
-                const testError = "Error in creating a login url";
+                const testError = {
+                    errorCode: "create_login_url_error",
+                    errorMessage: "Error in creating a login url"
+                }
                 sinon.stub(AuthorizationCodeClient.prototype, "getAuthCodeUrl").resolves(testNavUrl);
 				sinon.stub(PopupHandler.prototype, "initiateAuthRequest").throws(testError);
 				sinon.stub(CryptoOps.prototype, "generatePkceCodes").resolves({
@@ -1006,8 +1079,14 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                         scopes: TEST_CONFIG.DEFAULT_SCOPES
                     });
                 } catch (e) {
-                    expect(window.sessionStorage).to.be.empty;
-                    expect(`${e}`).to.be.eq(testError);
+                    // Test that error was cached for telemetry purposes and then thrown
+                    expect(window.sessionStorage).to.be.length(1);
+                    const failures = window.sessionStorage.getItem(`server-telemetry-${TEST_CONFIG.MSAL_CLIENT_ID}`);
+                    const failureObj = JSON.parse(failures) as ServerTelemetryCacheValue;
+                    expect(failureObj.failedRequests).to.be.length(2);
+                    expect(failureObj.failedRequests[0]).to.eq(ApiId.acquireTokenPopup);
+                    expect(failureObj.errors[0]).to.eq(testError.errorCode);
+                    expect(e).to.be.eq(testError);
                 }
             });
         });
@@ -1208,7 +1287,10 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
         });
 
         it("throws error that SilentFlowClient.acquireToken() throws", async () => {
-            const testError = "Error in creating a login url";
+            const testError = {
+                errorCode: "create_login_url_error",
+                errorMessage: "Error in creating a login url"
+            }
             const testAccount: AccountInfo = {
                 homeAccountId: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
                 environment: "login.windows.net",
@@ -1222,8 +1304,14 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                     account: testAccount
                 });
             } catch (e) {
-                expect(`${e}`).to.contain(testError);
-                expect(window.sessionStorage).to.be.empty;
+                // Test that error was cached for telemetry purposes and then thrown
+                expect(window.sessionStorage).to.be.length(1);
+                const failures = window.sessionStorage.getItem(`server-telemetry-${TEST_CONFIG.MSAL_CLIENT_ID}`);
+                const failureObj = JSON.parse(failures) as ServerTelemetryCacheValue;
+                expect(failureObj.failedRequests).to.be.length(2);
+                expect(failureObj.failedRequests[0]).to.eq(ApiId.acquireTokenSilent_silentFlow);
+                expect(failureObj.errors[0]).to.eq(testError.errorCode);
+                expect(e).to.be.eq(testError);
             }
         });
 
@@ -1368,10 +1456,10 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             expect(account).to.be.length(2);
         });
 
-        it("getAllAccounts returns null if no accounts signed in", () => {
+        it("getAllAccounts returns empty array if no accounts signed in", () => {
             window.sessionStorage.clear();
-            const account = pca.getAllAccounts();
-            expect(account).to.be.null;
+            const accounts = pca.getAllAccounts();
+            expect(accounts).to.deep.eq([]);
         });
 
         it("getAccountByUsername returns account specified", () => {
@@ -1393,5 +1481,10 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             expect(account).to.be.null;
         });
 
+        it("getAccountByUsername returns null if passed username is null", () => {
+            window.sessionStorage.clear();
+            const account = pca.getAccountByUsername(null);
+            expect(account).to.be.null;
+        });
     });
 });